# backend/app/api/routes.py
"""API routes — generated from the APS OpenAPI design (docs/openapi.json).

Each endpoint's method + path + params are real; the body is a TODO stub.
Generated by Autonomous Product Studio (APS) - fill in the TODOs to ship.
"""
from __future__ import annotations

from fastapi import APIRouter, HTTPException

router = APIRouter()


@router.get("/users")
def get_users():
    """List Users."""
    # TODO: implement GET /users
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.post("/users")
def post_users():
    """Create User."""
    # TODO: implement POST /users
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/users/{id}")
def get_users_x(id: str):
    """Get User."""
    # TODO: implement GET /users/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.put("/users/{id}")
def put_users(id: str):
    """Update User."""
    # TODO: implement PUT /users/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.delete("/users/{id}")
def delete_users(id: str):
    """Delete User."""
    # TODO: implement DELETE /users/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/patents")
def get_patents():
    """List Patents."""
    # TODO: implement GET /patents
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.post("/patents")
def post_patents():
    """Create Patent."""
    # TODO: implement POST /patents
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/patents/{id}")
def get_patents_x(id: str):
    """Get Patent."""
    # TODO: implement GET /patents/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.put("/patents/{id}")
def put_patents(id: str):
    """Update Patent."""
    # TODO: implement PUT /patents/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.delete("/patents/{id}")
def delete_patents(id: str):
    """Delete Patent."""
    # TODO: implement DELETE /patents/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/registrations")
def get_registrations():
    """List Registrations."""
    # TODO: implement GET /registrations
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.post("/registrations")
def post_registrations():
    """Create Registration."""
    # TODO: implement POST /registrations
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/registrations/{id}")
def get_registrations_x(id: str):
    """Get Registration."""
    # TODO: implement GET /registrations/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.put("/registrations/{id}")
def put_registrations(id: str):
    """Update Registration."""
    # TODO: implement PUT /registrations/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.delete("/registrations/{id}")
def delete_registrations(id: str):
    """Delete Registration."""
    # TODO: implement DELETE /registrations/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/forms")
def get_forms():
    """List Forms."""
    # TODO: implement GET /forms
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.post("/forms")
def post_forms():
    """Create Form."""
    # TODO: implement POST /forms
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/forms/{id}")
def get_forms_x(id: str):
    """Get Form."""
    # TODO: implement GET /forms/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.put("/forms/{id}")
def put_forms(id: str):
    """Update Form."""
    # TODO: implement PUT /forms/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.delete("/forms/{id}")
def delete_forms(id: str):
    """Delete Form."""
    # TODO: implement DELETE /forms/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")